+++ utils/CharEncoding.h
@@ -113,11 +113,11 @@ uint32_t utf8ToUtf32(const uint8_t *data, uint32_t len, utf32_t *bufOut, uint32_
 
 // 将 utf-32 的 code 转换为 utf-8 编码, @bufOut 需要至少 4 个字节的长度
 // 返回填入 bufOut 的字节长度
-uint32_t utf32CodeToUtf8Length(uint16_t code);
-uint32_t utf32CodeToUtf8(uint16_t code, uint8_t *bufOut);
-void utf32CodeToUtf8(uint16_t code, string &out);
+uint32_t utf32CodeToUtf8Length(uint32_t code);
+uint32_t utf32CodeToUtf8(uint32_t code, uint8_t *bufOut);
+void utf32CodeToUtf8(uint32_t code, string &out);
 
-uint32_t utf32CodeToUtf16Length(uint16_t code);
+uint32_t utf32CodeToUtf16Length(uint32_t code);
 
 // Big Endian to Little Endian, or vice versa
 void ucs2EncodingReverse(WCHAR *str, uint32_t nLen);
